#12 events moved to os.smog.events
