+++ src/OS.Smog.Events/OS.Smog.Events.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net462</TargetFramework>
+    <OutputType>library</OutputType>
+    <RuntimeIdentifier>win7-x86</RuntimeIdentifier>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+    <PlatformTarget>x86</PlatformTarget>
+  </PropertyGroup>
+
+
+  <ItemGroup>
+    <ProjectReference Include="..\OS.Events\OS.Events.csproj" />
+    <ProjectReference Include="..\OS.Smog.Dto\OS.Smog.Dto.csproj" />
+  </ItemGroup>
+
+</Project>
